+++ index.html
@@ -6,8 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://unpkg.com/sakura.css/css/sakura.css" type="text/css">
-    <script src="https://unpkg.com/@babel/standalone@7/babel.min.js"></script>
-    <script type="text/babel" data-type="module" data-presets="react" src="./unicode-to-latex.jsx"></script>
+    <script type="module" src="./unicode-to-latex.js"></script>
 </head>
 
 <body>
